+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.0.2 (2018/09/24)
 
 - rename `Singleton.instance` to `Singleton._instance`
+- add `SubclassableSingleton` class
 
 ## 0.0.1 (2018/09/24)